Improvements to RPN_EVAL  - Bug fixes  - Documentation fixes  - More documentation  - Conditional statements in RPN_EVAL  - Stack and global variables are now global to the scripts
I would prefer if the stack and global variables were preserved across executions of a script,
and local variables are local to that execution of the script (not just the line).

Even better would be variables shared between concurrently running scripts (1 instance per instance of LPHK).  that may be next.