A far more robust implementation of variables for the Mouse commands.   * only the mouse commands are implemented   * this is a stepping stone to a far more elegant solution where commands can validate themselves.   * updated documentation showing how to add new commands using variables   * variables are readable only at present in commands other than RPN_EVAL
try the following examples

M_MOVE 1.2 7   

This will fail because the first parameter is not an integer.  negative numbers also cause failure to validate.

RPN_EVAL 1.5 > a
RPN_EVAL -1 > b
M_MOVE a b

This validates correctly, but fails at run-time an error indicating the second parameter is negative.  The first parameter is converted to an integer, resulting in the value 1 being used.  Note that this is currently a simple truncation.