Faster loading for layouts.  If layouts have a lot of subroutines, or the scripts in the buttons are huge, it can take a considerable amount of time to load them.
An optional keyword has been added to the LAOAD_LAYOUT command to control the amount of validation and subroutine magic.  The parameter "NORMAL" (default) does the normal thing, unloading subroutines, reloading the ones in the layout, loading the layout and checking both the subroutines and buttons.  The keyword "FAST" retains the existing subroutines and does not load any specified in the new layout.  However the loaded buttons are checked for correctness (and will fail at this if they call a subroutine that is not loaded.  To use this successfully, you should ensure that all of your layouts have the same set of subroutines loaded within them.  The final option "UNCHECKED" does the same as "FAST", but does NOT check buttons for correctness.  Unchecked is significantly faster than "FAST" where there are a lot of buttons, or the buttons have long scripts.py.  With "UNCHECKED" parts of the script are still checked.  The parts that are checked are the headers since these are the lines that set titles and help strings.